refactor: textarea 로직 분리 전 Board 내 유틸 함수 분리
